Move fused modules
